downloader: refactor
From my point of view it makes code a bit easier to understand
and hides the downloader implementation, but FetchPosts
is still is a bit of a mess.

Benchmark showed that this also increased performance a bit.

Now you can get all the errors that happened during
execution in Stats.Error, which also will be nicely formatted.