saver: close the stream on a separate goroutine
This unblocks the main goroutine from continuing and lets the
program finish earlier. Previously, the program would wait a
non-insignificant amount of time for all the workers to end
before closing, making the user wait for pretty much no reason.
In most cases, the program will end before the stream closes,
and in cases where it doesn't happen, this approach will still
close the stream correctly in the background and won't leak goroutines.